# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DukaanBazar Indian General Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  pincode?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfileUpdate {
  name?: string;
  phone?: string;
  address?: string;
  city?: string;
  state?: string;
  pincode?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount: number;
}

export interface CategoryInput {
  name: string;
  description?: string;
  imageUrl?: string;
}

export interface CategoryUpdate {
  name?: string;
  description?: string;
  imageUrl?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  categoryId: number;
  categoryName: string;
  /** @nullable */
  brand?: string | null;
  stock: number;
  imageUrl: string;
  images?: string[];
  rating: number;
  reviewCount: number;
  isFeatured?: boolean;
  tags?: string[];
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  categoryId: number;
  brand?: string;
  stock: number;
  imageUrl: string;
  images?: string[];
  isFeatured?: boolean;
  tags?: string[];
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  categoryId?: number;
  brand?: string;
  stock?: number;
  imageUrl?: string;
  isFeatured?: boolean;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  imageUrl: string;
  stock: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface WishlistItem {
  id: number;
  productId: number;
  productName: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  imageUrl: string;
  stock: number;
  rating?: number;
  addedAt: string;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  imageUrl: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  status: OrderStatus;
  total: number;
  subtotal?: number;
  shippingFee?: number;
  paymentMethod?: string;
  paymentStatus?: string;
  items: OrderItem[];
  shippingAddress: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  pincode?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  shippingAddress: string;
  city?: string;
  state?: string;
  pincode?: string;
  paymentMethod: string;
  notes?: string;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface Review {
  id: number;
  productId: number;
  userId: number;
  userName: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
}

export interface CategoryStat {
  name: string;
  productCount: number;
  revenue: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface MonthRevenue {
  month: string;
  revenue: number;
  orders: number;
}

export interface AdminStats {
  totalUsers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  recentOrders: Order[];
  topCategories: CategoryStat[];
  ordersByStatus: StatusCount[];
  revenueByMonth: MonthRevenue[];
}

export type ListProductsParams = {
  page?: number;
  limit?: number;
  categoryId?: number;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  sortBy?: string;
  inStock?: boolean;
};

export type ListAllOrdersParams = {
  page?: number;
  status?: string;
};